new database template
permission are checked for all operations except 'visible'.
You can login as:
admin -> admin
autor -> autor
lerner ->lerner
gast -> gast


git-svn-id: http://svn.ilias.de/svn/ilias/trunk@44 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ classes/class.Admin.php
@@ -31,12 +31,21 @@ class Admin
 	function deleteObject()
 	{
 		global $tree;
-
+		
 		$rbacadmin = new RbacAdminH($this->ilias->db);
+		$rbacsystem = new RbacSystemH($this->ilias->db);
 		foreach($_POST["id"] as $id)
 		{
-			$tree->deleteTree($id);
-			$rbacadmin->revokePermission($id);
+			// CHECK ACCESS
+			if($rbacsystem->checkAccess('delete',$id,$_GET["obj_id"]))
+			{
+				$tree->deleteTree($id);
+				$rbacadmin->revokePermission($id);
+			}
+			else
+			{
+				$_SESSION["Error_Message"] = "No permission to delete Object";
+			}
 		}
 		header("Location: content.php?obj_id=$_GET[obj_id]&parent=$_GET[parent]");
 	}
